requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.5
fpdf==1.7.2
h11==0.14.0
idna==3.10
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.3
psycopg==3.2.3
pydantic==2.10.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
